Move EventHandler to own file
